requirements.txt
requests
beautifulsoup4
python-docx
html2docx
